refactor(cli): get default clustermgr from config
